+++ lib/src/codecs/codecs.dart
@@ -3,3 +3,4 @@ export 'string/string_codec.dart';
 export 'nmtoken/nmtoken_codec.dart';
 export 'normalized_string/normalized_string_codec.dart';
 export 'name/name_codec.dart';
+export 'token/token_codec.dart';