+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Implement `xsd:nmtoken` data type codec
 - Implement `xsd:normalizedString` data type codec
 - Implement `xsd:Name` data type codec
+- Implement `xsd:token` data type codec